Started making changes to expand the site
I'm making some changes to a few pages and sections to make way for
expanding the site! Guides are moved to a folder caled "fiesta" and are
now called "Fiesta Guides" I'm planning on keeping this change and also
creating an entrt for Final Fantasy XIV, but other guides will be in a
new guides section.